+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 [0.2.0] 01.04.2026
 
+- [IGS-22] - Upload photo to storage
 - [IGS-14] - Auth service, supabase connect